+++ CHANGELOG.md
@@ -102,6 +102,7 @@
 ## mini.pick
 
 - BREAKING: rename `prompt_cursor` in `config.window` to `prompt_caret` for better naming consistency. It works for now, but will stop in the next release. Sorry for the inconvenience.
+- FEATURE: add `MiniPickPromptCaret` and `MiniPickPromptPrefix` highlight groups to allow finer customization of picker's prompt.
 
 ## mini.statusline